fix: fixed the biggest bug in the fucking project that was hunting me from the start

+++ Kernel/Memory/memoryMap.h
@@ -82,6 +82,6 @@ void initializeFrames(void);
 void settingReservedFrames(void);
 int allocateFrame(unsigned long long address);
 int freeFrame(unsigned long long address);
-int allocateFreeFrame(void);
+unsigned long long allocateFreeFrame(void);
 int allocateFreeUserFrame(void);
 #endif
